beginning of tilemap for water.
Tilemap Collider 2D and NavMesh Obstacle inside of WaterTile layer. I am wondering if the navmesh really is making the whole map a collider (and navmesh agent wouldn't be able to be in it at all) or if it is  acting on just the water tiles the same way the 2d collider is.
This is a square tileset, but I see no problem using it in ISO. Let me know your thoughts