more map fill in
Should have the map totally filled out By end of this sprint.
Next I'd like to put in a forest over-layer that where the over-layer has no gaps, the armies should be able to "hide". However, since the lamanite armies are nav agents, this will only challenge the nephite armies